backend/requirements.txt
# ── Core Framework ───────────────────────────────────────
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic==2.10.*
pydantic-settings==2.7.*

# ── Supabase ─────────────────────────────────────────────
supabase==2.*

# ── Security ─────────────────────────────────────────────
cryptography==44.*
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*
bcrypt==4.*

# ── LLM & Agent ─────────────────────────────────────────
langchain-core==0.3.*
langgraph==0.3.*
langchain-google-genai==2.*
# Optional providers (install as needed):
# langchain-openai==0.3.*
# langchain-groq==0.3.*

# ── HTTP Client ──────────────────────────────────────────
httpx==0.28.*

# ── Embedding / RAG (Phase 3) ───────────────────────────
# langchain-community==0.3.*
# pypdf==5.*
# python-pptx==1.*
